update: fixing file naming user management
