+++ src/pages/UserManagementPages/role/index.tsx
@@ -1,26 +1,34 @@
-import PageBreadcrumb from "../../components/common/PageBreadCrumb";
-import ComponentCard from "../../components/common/ComponentCard";
-import PageMeta from "../../components/common/PageMeta";
-import SearchBar from "../../components/atoms/SearchBar";
-import Pagination from "../../components/atoms/Pagination";
-import useFetchUserData from "../../hooks/userManagement/useFetchUsers";
-import useFetchUserDelete from "../../hooks/userManagement/useFetchUserDelete";
-import HandleDeleteData from "../../services/utils/handleDeleteData";
+import PageBreadcrumb from "../../../components/common/PageBreadCrumb";
+import ComponentCard from "../../../components/common/ComponentCard";
+import PageMeta from "../../../components/common/PageMeta";
+import SearchBar from "../../../components/atoms/SearchBar";
+import Pagination from "../../../components/atoms/Pagination";
+import useFetchRoleData from "../../../hooks/userManagement/useFetchRoles";
+import useFetchRoleDelete from "../../../hooks/userManagement/useFetchRoleDelete";
+import HandleDeleteData from "../../../services/utils/handleDeleteData";
 import { useSelector } from "react-redux";
-import { StateContext } from "../../types/app.type";
-import { userManagementStateContext } from "../../types/app.type";
-import { UserData } from "../../types/userManagement.type";
+import { StateContext } from "../../../types/app.type";
+import { userManagementStateContext } from "../../../types/app.type";
+import { RoleData } from "../../../types/userManagement.type";
 import { Link } from "react-router";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 
-export default function UserManagement() {
-  const [searchValue, setSearchValue] = useState('');
-  const [currentPage, setCurrentPage] = useState(1);
-  const [idToDetele, setIdToDelete] = useState<number>();
-  useFetchUserData(String(currentPage), searchValue);
-  useFetchUserDelete(idToDetele);
+export default function RoleManagement() {
   const userManagementData: userManagementStateContext  = useSelector((state:StateContext) => state.userManagement);
-  console.log(userManagementData.userData)
+  const [searchValue, setSearchValue] = useState('');
+  const [currentPage, setCurrentPage] = useState(userManagementData.roleDataMeta.current_page || 1);
+  const [idToDetele, setIdToDelete] = useState<number>();
+  const ReloadData = () => {
+    useFetchRoleData(String(currentPage), searchValue); 
+  };
+  
+  useFetchRoleDelete(idToDetele,  ReloadData);
+  useFetchRoleData(String(currentPage), searchValue);
+
+  useEffect(()=>{
+    setCurrentPage(1)
+  },[searchValue])
+
   return (
     <>
       <PageMeta
@@ -36,7 +44,7 @@ export default function UserManagement() {
                     <SearchBar searchValue={searchValue} setSearchValue={setSearchValue} />
                 </div>
                 <div className="content-center bg-white dark:bg-gray-900">
-                  <Link to={'/user-management/create'}  className="px-8 py-3 rounded-md text-white bg-green-600 hover:bg-green-700">
+                  <Link to={'/role-management/create'}  className="px-8 py-3 rounded-md text-white bg-green-600 hover:bg-green-700">
                     Create
                   </Link>
                 </div>
@@ -45,10 +53,7 @@ export default function UserManagement() {
                   <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
                       <tr>
                           <th scope="col" className="px-6 py-3">
-                              Name
-                          </th>
-                          <th scope="col" className="px-6 py-3">
-                              Email
+                              Role name
                           </th>
                           <th scope="col" className="px-6 py-3">
                               Action
@@ -56,23 +61,20 @@ export default function UserManagement() {
                       </tr>
                   </thead>
                   <tbody>
-                    {userManagementData.userData.map((item: UserData ) => (
+                    {(searchValue ? userManagementData.roleData.searchResults[searchValue]?.[currentPage] : userManagementData.roleData.pages[currentPage])?.map((item: RoleData ) => (
                       <tr key={item.id} className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 border-gray-200 hover:bg-gray-50 dark:hover:bg-gray-600">
                         <th scope="row" className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                            {item.name}
+                            {item.nama}
                         </th>
                         <th scope="row" className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                            {item.email}
-                        </th>
-                        <th scope="row" className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white">
-                          <Link to={`/user-management/${item.id}`} className="font-medium text-blue-600 dark:text-blue-500 hover:underline">Edit</Link>   
-                          <span onClick={() => {HandleDeleteData(item.id, {name: item.name, kind: "User"}, setIdToDelete) }} className="font-medium text-red-600 dark:text-red-500 hover:underline ms-3">Delete</span>
+                          <Link to={`/role-management/${item.id}`} className="font-medium text-blue-600 dark:text-blue-500 hover:underline">Edit</Link>   
+                          <span onClick={() => {HandleDeleteData(item.id, {name: item.nama, kind: "Role"}, setIdToDelete) }} className="font-medium text-red-600 dark:text-red-500 hover:underline ms-3">Delete</span>
                         </th>
                     </tr>
                     ))}
                   </tbody>
               </table>
-              <Pagination currentPage={currentPage} setCurrentPage={setCurrentPage} data={userManagementData.userDataMeta}/>
+              <Pagination currentPage={currentPage} setCurrentPage={setCurrentPage} data={userManagementData.roleDataMeta}/>
           </div>
         </ComponentCard>
       </div>
