+++ src/pages/UserManagementPages/user/create.tsx
@@ -1,17 +1,17 @@
-import PageBreadcrumb from "../../components/common/PageBreadCrumb";
-import ComponentCard from "../../components/common/ComponentCard";
-import PageMeta from "../../components/common/PageMeta";
-import Pagination from "../../components/atoms/Pagination";
-import useFetchUserCreate from "../../hooks/userManagement/useFetchUserCreate";
-import Label from "../../components/form/Label";
-import { EyeIcon, EyeCloseIcon } from "../../icons";
+import PageBreadcrumb from "../../../components/common/PageBreadCrumb";
+import ComponentCard from "../../../components/common/ComponentCard";
+import PageMeta from "../../../components/common/PageMeta";
+import Pagination from "../../../components/atoms/Pagination";
+import useFetchUserCreate from "../../../hooks/userManagement/useFetchUserCreate";
+import Label from "../../../components/form/Label";
+import { EyeIcon, EyeCloseIcon } from "../../../icons";
 import { useSelector } from "react-redux";
-import { StateContext } from "../../types/app.type";
-import { userManagementStateContext } from "../../types/app.type";
-import { UserForm } from "../../types/form.type";
+import { StateContext } from "../../../types/app.type";
+import { userManagementStateContext } from "../../../types/app.type";
+import { UserForm } from "../../../types/form.type";
 import { useForm } from "react-hook-form";
 import { useState } from "react";
-import useFetchRoleData from "../../hooks/userManagement/useFetchRoles";
+import useFetchRoleData from "../../../hooks/userManagement/useFetchRoles";
 
 export default function CreateUser() {
