+++ src/services/routes.tsx
@@ -5,12 +5,12 @@ import SignIn from "../pages/AuthPages/SignIn";
 import { ScrollToTop } from "../components/common/ScrollToTop";
 import AppLayout from "../layout/AppLayout";
 import Home from "../pages/Dashboard/Home";
-import UserManagement from "../pages/UserManagementPages/UserManagements";
-import RoleManagement from "../pages/UserManagementPages/RoleManagements";
-import EditRole from "../pages/UserManagementPages/RoleEdit";
-import EditUser from "../pages/UserManagementPages/UserEdit";
-import CreateRole from "../pages/UserManagementPages/RoleCreate";
-import CreateUser from "../pages/UserManagementPages/UserCreate";
+import UserManagement from "../pages/UserManagementPages/user";
+import RoleManagement from "../pages/UserManagementPages/role";
+import EditRole from "../pages/UserManagementPages/role/edit";
+import EditUser from "../pages/UserManagementPages/user/edit";
+import CreateRole from "../pages/UserManagementPages/role/create";
+import CreateUser from "../pages/UserManagementPages/user/create";
 import NotFound from "../pages/OtherPage/NotFound";
 import CategoryManagement from "../pages/ArtikelManagementPages/CategoryManagements";
 import CreateCategory from "../pages/ArtikelManagementPages/CategoryCreate";
